HHH-15288 Switch to SMALLINT as recommended type for enums

+++ hibernate-core/src/test/java/org/hibernate/orm/test/mapping/SmokeTests.java
@@ -96,7 +96,7 @@ public class SmokeTests {
 
 			assertThat(
 					genderAttrMapping.getJdbcMapping().getJdbcType(),
-					is( jdbcTypeRegistry.getDescriptor( Types.TINYINT ) )
+					is( jdbcTypeRegistry.getDescriptor( Types.SMALLINT ) )
 			);
 		}
